@@ -3,6 +3,6 @@ import { Course } from "./course";
 export interface CoursePage {
   courses: Course[];
   totalElements: number;
-  totalPages: number;
+  totalPages?: number;
 
 }
