@@ -1,24 +1,19 @@
 import { Injectable } from '@angular/core';
-import {
-  Router, Resolve,
-  RouterStateSnapshot,
-  ActivatedRouteSnapshot
-} from '@angular/router';
+import { ActivatedRouteSnapshot, RouterStateSnapshot } from '@angular/router';
 import { Observable, of } from 'rxjs';
+
 import { CoursesService } from '../services/courses.service';
 import { Course } from '../models/course';
 
 @Injectable({
   providedIn: 'root'
 })
-export class CourseResolver implements Resolve<Course> {
-  constructor(private service: CoursesService) {
+export class CourseResolver {
+  constructor(private service: CoursesService) { }
 
-  }
   resolve(route: ActivatedRouteSnapshot, state: RouterStateSnapshot): Observable<Course> {
-
     if (route.params && route.params['id']) {
-      return this.service.findById(route.params['id'])
+      return this.service.findById(route.params['id']);
     }
 
     return of({ _id: '', name: '', category: '', lessons: [] });
