@@ -8,8 +8,8 @@ describe('ConfirmationDialogComponent', () => {
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      declarations: [ ConfirmationDialogComponent ]
-    })
+    imports: [ConfirmationDialogComponent]
+})
     .compileComponents();
 
     fixture = TestBed.createComponent(ConfirmationDialogComponent);
