@@ -1,4 +1,5 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
+import { MAT_DIALOG_DATA, MatDialogModule, MatDialogRef } from '@angular/material/dialog';
 
 import { ErrorDialogComponent } from './error-dialog.component';
 
@@ -8,12 +9,13 @@ describe('ErrorDialogComponent', () => {
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      declarations: [ ErrorDialogComponent ]
-    })
-    .compileComponents();
-  });
+      imports: [MatDialogModule, ErrorDialogComponent],
+      providers: [
+        { provide: MAT_DIALOG_DATA, useValue: 'Error' },
+        { provide: MatDialogRef, useValue: {} }
+      ]
+    }).compileComponents();
 
-  beforeEach(() => {
     fixture = TestBed.createComponent(ErrorDialogComponent);
     component = fixture.componentInstance;
     fixture.detectChanges();
@@ -21,5 +23,15 @@ describe('ErrorDialogComponent', () => {
 
   it('should create', () => {
     expect(component).toBeTruthy();
+  });
+
+  it('should launch an error dialog with a message and a Close button', () => {
+    const errorMessageDom = fixture.nativeElement.querySelector(
+      '.mat-mdc-dialog-content'
+    );
+    expect(errorMessageDom.textContent).toContain('Error');
+
+    const okBtn = fixture.nativeElement.querySelector('button');
+    expect(okBtn.textContent).toContain('Close');
   });
 });
